refactor: Merge `section` into `fetch`.
