+++ imap-codec/benches/serialize_command.rs
@@ -4,8 +4,7 @@ use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use imap_codec::{
     codec::Encode,
     command::{Command, CommandBody},
-    fetch::{MacroOrMessageDataItemNames, MessageDataItemName},
-    section::Section,
+    fetch::{MacroOrMessageDataItemNames, MessageDataItemName, Section},
 };
 
 fn criterion_benchmark(c: &mut Criterion) {
