+++ imap-codec/src/fetch.rs
@@ -16,11 +16,10 @@ use nom::{
 use crate::{
     body::body,
     codec::IMAPResult,
-    core::{nstring, number, nz_number, NonEmptyVec},
+    core::{astring, nstring, number, nz_number, AString, NonEmptyVec},
     datetime::date_time,
     envelope::envelope,
     flag::flag_fetch,
-    section::section,
 };
 
 /// `fetch-att = "ENVELOPE" /
@@ -187,6 +186,99 @@ pub(crate) fn uniqueid(input: &[u8]) -> IMAPResult<&[u8], NonZeroU32> {
     nz_number(input)
 }
 
+/// `section = "[" [section-spec] "]"`
+pub(crate) fn section(input: &[u8]) -> IMAPResult<&[u8], Option<Section>> {
+    delimited(tag(b"["), opt(section_spec), tag(b"]"))(input)
+}
+
+/// `section-spec = section-msgtext / (section-part ["." section-text])`
+pub(crate) fn section_spec(input: &[u8]) -> IMAPResult<&[u8], Section> {
+    alt((
+        map(section_msgtext, |part_specifier| match part_specifier {
+            PartSpecifier::PartNumber(_) => unreachable!(),
+            PartSpecifier::Header => Section::Header(None),
+            PartSpecifier::HeaderFields(fields) => Section::HeaderFields(None, fields),
+            PartSpecifier::HeaderFieldsNot(fields) => Section::HeaderFieldsNot(None, fields),
+            PartSpecifier::Text => Section::Text(None),
+            PartSpecifier::Mime => unreachable!(),
+        }),
+        map(
+            tuple((section_part, opt(tuple((tag(b"."), section_text))))),
+            |(part_number, maybe_part_specifier)| {
+                if let Some((_, part_specifier)) = maybe_part_specifier {
+                    match part_specifier {
+                        PartSpecifier::PartNumber(_) => unreachable!(),
+                        PartSpecifier::Header => Section::Header(Some(Part(part_number))),
+                        PartSpecifier::HeaderFields(fields) => {
+                            Section::HeaderFields(Some(Part(part_number)), fields)
+                        }
+                        PartSpecifier::HeaderFieldsNot(fields) => {
+                            Section::HeaderFieldsNot(Some(Part(part_number)), fields)
+                        }
+                        PartSpecifier::Text => Section::Text(Some(Part(part_number))),
+                        PartSpecifier::Mime => Section::Mime(Part(part_number)),
+                    }
+                } else {
+                    Section::Part(Part(part_number))
+                }
+            },
+        ),
+    ))(input)
+}
+
+/// `section-msgtext = "HEADER" / "HEADER.FIELDS" [".NOT"] SP header-list / "TEXT"`
+///
+/// Top-level or MESSAGE/RFC822 part
+pub(crate) fn section_msgtext(input: &[u8]) -> IMAPResult<&[u8], PartSpecifier> {
+    alt((
+        map(
+            tuple((tag_no_case(b"HEADER.FIELDS.NOT"), sp, header_list)),
+            |(_, _, header_list)| PartSpecifier::HeaderFieldsNot(header_list),
+        ),
+        map(
+            tuple((tag_no_case(b"HEADER.FIELDS"), sp, header_list)),
+            |(_, _, header_list)| PartSpecifier::HeaderFields(header_list),
+        ),
+        value(PartSpecifier::Header, tag_no_case(b"HEADER")),
+        value(PartSpecifier::Text, tag_no_case(b"TEXT")),
+    ))(input)
+}
+
+#[inline]
+/// `section-part = nz-number *("." nz-number)`
+///
+/// Body part nesting
+pub(crate) fn section_part(input: &[u8]) -> IMAPResult<&[u8], NonEmptyVec<NonZeroU32>> {
+    map(
+        separated_list1(tag(b"."), nz_number),
+        NonEmptyVec::unvalidated,
+    )(input)
+}
+
+/// `section-text = section-msgtext / "MIME"`
+///
+/// Text other than actual body part (headers, etc.)
+pub(crate) fn section_text(input: &[u8]) -> IMAPResult<&[u8], PartSpecifier> {
+    alt((
+        section_msgtext,
+        value(PartSpecifier::Mime, tag_no_case(b"MIME")),
+    ))(input)
+}
+
+/// `header-list = "(" header-fld-name *(SP header-fld-name) ")"`
+pub(crate) fn header_list(input: &[u8]) -> IMAPResult<&[u8], NonEmptyVec<AString>> {
+    map(
+        delimited(tag(b"("), separated_list1(sp, header_fld_name), tag(b")")),
+        NonEmptyVec::unvalidated,
+    )(input)
+}
+
+#[inline]
+/// `header-fld-name = astring`
+pub(crate) fn header_fld_name(input: &[u8]) -> IMAPResult<&[u8], AString> {
+    astring(input)
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -327,6 +419,60 @@ mod tests {
 
         for test in tests {
             known_answer_test_encode(test);
+        }
+    }
+
+    #[test]
+    fn test_encode_section() {
+        let tests = [
+            (
+                Section::Part(Part(NonEmptyVec::from(NonZeroU32::try_from(1).unwrap()))),
+                b"1".as_ref(),
+            ),
+            (Section::Header(None), b"HEADER"),
+            (
+                Section::Header(Some(Part(NonEmptyVec::from(
+                    NonZeroU32::try_from(1).unwrap(),
+                )))),
+                b"1.HEADER",
+            ),
+            (
+                Section::HeaderFields(None, NonEmptyVec::from(AString::try_from("").unwrap())),
+                b"HEADER.FIELDS (\"\")",
+            ),
+            (
+                Section::HeaderFields(
+                    Some(Part(NonEmptyVec::from(NonZeroU32::try_from(1).unwrap()))),
+                    NonEmptyVec::from(AString::try_from("").unwrap()),
+                ),
+                b"1.HEADER.FIELDS (\"\")",
+            ),
+            (
+                Section::HeaderFieldsNot(None, NonEmptyVec::from(AString::try_from("").unwrap())),
+                b"HEADER.FIELDS.NOT (\"\")",
+            ),
+            (
+                Section::HeaderFieldsNot(
+                    Some(Part(NonEmptyVec::from(NonZeroU32::try_from(1).unwrap()))),
+                    NonEmptyVec::from(AString::try_from("").unwrap()),
+                ),
+                b"1.HEADER.FIELDS.NOT (\"\")",
+            ),
+            (Section::Text(None), b"TEXT"),
+            (
+                Section::Text(Some(Part(NonEmptyVec::from(
+                    NonZeroU32::try_from(1).unwrap(),
+                )))),
+                b"1.TEXT",
+            ),
+            (
+                Section::Mime(Part(NonEmptyVec::from(NonZeroU32::try_from(1).unwrap()))),
+                b"1.MIME",
+            ),
+        ];
+
+        for test in tests {
+            known_answer_test_encode(test)
         }
     }
 }
